requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
torchmetrics>=1.3.0

# Data handling
pandas>=2.1.0
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.4.0

# Augmentation
albumentations>=1.3.0

# Training utilities
tensorboard>=2.15.0
tqdm>=4.66.0
pyyaml>=6.0.1

# API serving
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Utilities
matplotlib>=3.8.0
seaborn>=0.13.0
